refactor: Remove components/button.tsx
